server/requirements.txt
django
django-ninja
jupyter
langchain
langchain-openai
langgraph
langgraph-supervisor
python-decouple
permit
requests
streamlit
psycopg[binary]
